FT: ListMetrics for bucket resource
This commit introduces route to retrieve metrics for a given list
of buckets. It takes timeRange and a list of buckets from the request
body JSON and responds back the list of metrics. Since this data is
loosely coupled with S3 (it has no knowledge of a bucket's existence
in S3), it responds back with JSON containing 0s as metrics for the
missing bucket.