ft: use hiredis parser
hiredis is a native C module for redis-cli. Cosbench tests showed this module performed
marginally better than the default javascript parser. ioredis module uses hiredis if it's
installed or reverts to it's in-built javascript parser.
The dropBufferSupport option prefers strings over buffers for replies and hiredis has
better performance with this option enabled.
Module is being installed as an optional dependency to favor open source installs on
machines which may not have gcc / g++ / python installed.