bugfix: S3C-3620 don't raise exception on ioredis client error
When ioredis emits an error (e.g. connection issue), instead of
unconditionally raising it in the RedisClient wrapper, only raise it
if there is at least one listener for the 'error' event.

Some users of RedisClient (v2 cache and reindex tasks) do not set a
listener on error events, resulting in assertions being raised and
process crash.